requirements.txt
streamlit
pandas
plotly
numpy
PyGithub